requirements.txt
flask 
opencv-python 
scikit-image 
imutils 
pillow
